refactor: Use dataclasses instead of pydantic.BaseClass
Since dataclasses are the idiomatic Python way of defining the shape of classes,
we're effectively simplifying the setup by going back to basics.

Furthermore, the use of dataclasses eliminates the need for us to inherit from
pydantic.BaseModel which caused some type checking issues with pytype. This
typing issue is rendered moot by moving away from the "inherit BaseModel"
design.